refactor: better abstract array

refactor: add class getitem


refactor: do document list


refactor: remove unsceray fodler


refactor: remove old array
