--- docarray/array/mixins/__init__.py
@@ -1,4 +1,0 @@
-from docarray.array.mixins.attribute import GetAttributeArrayMixin
-from docarray.array.mixins.proto import ProtoArrayMixin
-
-__all__ = ['ProtoArrayMixin', 'GetAttributeArrayMixin']
